+++ src/Helpers/Logger.cs
@@ -39,7 +39,6 @@ public static class Logger
         if (ex != null)
         {
             Log(ex.ToString());
-            Telemetry.TrackException(ex);
         }
     }